fix IINC, LINC bug

+++ sperl_vm.c
@@ -908,11 +908,11 @@ void SPerl_VM_call_sub(SPerl* sperl, SPerl_VM* vm, const char* sub_base_name) {
         pc++;
         continue;
       case SPerl_BYTECODE_C_CODE_IINC:
-        *(int32_t*)&vars[bytecodes[pc + 1]] += bytecodes[pc + 2];
+        *(int32_t*)&vars[bytecodes[pc + 1]] += (int8_t)bytecodes[pc + 2];
         pc += 3;
         continue;
       case SPerl_BYTECODE_C_CODE_LINC:
-        vars[bytecodes[pc + 1]] += bytecodes[pc + 2];
+        vars[bytecodes[pc + 1]] += (int8_t)bytecodes[pc + 2];
         pc += 3;
         continue;
       case SPerl_BYTECODE_C_CODE_I2L:
@@ -1522,11 +1522,11 @@ void SPerl_VM_call_sub(SPerl* sperl, SPerl_VM* vm, const char* sub_base_name) {
           pc +=4;
         }
         else if (bytecodes[pc + 1] == SPerl_BYTECODE_C_CODE_IINC) {
-          *(int32_t*)&vars[(bytecodes[pc + 2] << 8) + bytecodes[pc + 3]] += (bytecodes[pc + 4] << 8) + bytecodes[pc + 5];
+          *(int32_t*)&vars[(bytecodes[pc + 2] << 8) + bytecodes[pc + 3]] += (int8_t)((bytecodes[pc + 4] << 8) + bytecodes[pc + 5]);
           pc += 6;
         }
         else if (bytecodes[pc + 1] == SPerl_BYTECODE_C_CODE_LINC) {
-          vars[(bytecodes[pc + 2] << 8) + bytecodes[pc + 3]] += (bytecodes[pc + 4] << 8) + bytecodes[pc + 5];
+          vars[(bytecodes[pc + 2] << 8) + bytecodes[pc + 3]] += (int8_t)((bytecodes[pc + 4] << 8) + bytecodes[pc + 5]);
           pc += 6;
         }
         continue;
